Refactored duplicate code to base class. Works with departures also
